Support audio and video tags in epub-fetch
Currently, this is implemented in a naive way - by fetching media
resources into memory in their entirety as blobs.t

This has a downside of consuming lots of browser heap space.

However, if the resource is fetched from a remote ZIP file, streaming is
impossible due to absence of random access. In the future, we might
fetch media files to local filesystem using the experimental HTML5
temporary filesystem API and stream from over there.